Blindaje de dinero: tope de 10.000 EUR por importe y sumas en float8 (una tarea de 1M EUR x22 desbordaba el int y tumbaba la cuenta)

+++ src/app/api/cron/weekly-summary/route.ts
@@ -36,7 +36,7 @@ async function run(req: Request): Promise<Response> {
       pointsName: kids.pointsName,
       pointsIcon: kids.pointsIcon,
       n: sql<number>`count(${completions.id})::int`,
-      c: sql<number>`coalesce(sum(${completions.valueCents}),0)::int`,
+      c: sql<number>`coalesce(sum(${completions.valueCents}),0)::float8`,
     })
     .from(kids)
     .leftJoin(
